requirements.txt
psycopg2-binary==2.9.3
Django==4.0.4
django-environ==0.8.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
environ==1.0
django-filter==21.1
Pillow==9.1.0
ipdb==0.13.9
selenium==4.2.0
beautifulsoup4==4.11.1
requests==2.27.1
lxml==4.9.0
webdriver-manager==3.7.0
backports.zoneinfo==0.2.1
PyJWT==2.3.0
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
celery==5.2.7
flower==1.0.0
django-celery-beat==2.3.0
redis==4.3.3
django-elasticsearch==7.0.2
django-elasticsearch-dsl==7.2.2
django-elasticsearch-dsl-drf==0.22.4
elasticsearch==7.17.4
elasticsearch-dsl==7.4.0